Fix dynamic updates grid items, deduplicate crud printer form as component
